Move metadata to layout instead of page
