+++ frontend/app/(views)/(website)/lines/layout.tsx
@@ -1,6 +1,14 @@
 /* * */
 
 import { LinesListContextProvider } from '@/contexts/LinesList.context';
+import { type Metadata } from 'next';
+
+/* * */
+
+export const metadata: Metadata = {
+	description: 'Descubra as linhas de autocarro da CMetropolitana.',
+	title: 'CMetropolitana | Linhas',
+};
 
 /* * */
