+++ frontend/app/(views)/(website)/lines/page.tsx
@@ -1,14 +1,6 @@
 /* * */
 
 import { LinesList } from '@/components/lines/LinesList';
-import { Metadata } from 'next';
-
-/* * */
-
-export const metadata: Metadata = {
-	description: 'Descubra as linhas de autocarro da CMetropolitana.',
-	title: 'CMetropolitana | Linhas',
-};
 
 /* * */
