+++ frontend/app/(views)/(website)/stops/layout.tsx
@@ -1,6 +1,14 @@
 /* * */
 
 import { StopsListContextProvider } from '@/contexts/StopsList.context';
+import { type Metadata } from 'next';
+
+/* * */
+
+export const metadata: Metadata = {
+	description: 'Descubra as paragens de autocarro da CMetropolitana.',
+	title: 'CMetropolitana | Paragens',
+};
 
 /* * */
