+++ frontend/app/(views)/(website)/stops/page.tsx
@@ -1,14 +1,6 @@
 /* * */
 
 import { StopsList } from '@/components/stops/StopsList';
-import { Metadata } from 'next';
-
-/* * */
-
-export const metadata: Metadata = {
-	description: 'Descubra as paragens de autocarro da CMetropolitana.',
-	title: 'CMetropolitana | Paragens',
-};
 
 /* * */
